add state to TreeMapSection Component
